requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
SQLAlchemy==2.0.*
celery[redis]==5.3.*
pytesseract==0.3.*
Pillow==10.0.*
aiofiles==23.1.*
python-dotenv==1.0.*
pytest==7.0.*
python-multipart